# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "open-cam API",
    "description": "open-cam 是视频流分析与监控管理平台。\n\n- **摄像头**：接入局域网 RTSP 流或本地视频文件，YOLO（cuda/mps/cpu 自适应）检测与跟踪。\n- **规则引擎**：区域入侵 / 徘徊逗留 / 人数统计 / 区域人数 / 越线计数，支持生效时段与冷却去抖。\n- **事件**：命中规则即落库并存快照，可经 VLM（OpenAI 兼容接口）异步复核；处置闭环支持关注星标、负责人指派、状态流转与 webhook 通知，全程留痕。\n- **方案包**：行业规则模板包（连锁零售 / 美容美发 / 餐饮 / 快餐），一键安装与应用。\n- **统计**：分时段进出店客流等聚合视图。\n\n交互式调试见 [Swagger UI](/docs)，结构化阅读见 [ReDoc](/redoc)。\n",
    "contact": {
      "name": "open-cam",
      "url": "https://github.com/local/open-cam"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/videos": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "已上传视频列表",
        "operationId": "list_videos_api_videos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VideoOut"
                  },
                  "type": "array",
                  "title": "Response List Videos Api Videos Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "videos"
        ],
        "summary": "上传视频文件",
        "operationId": "upload_video_api_videos_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_video_api_videos_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/videos/{video_id}": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "视频详情",
        "operationId": "get_video_api_videos__video_id__get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "videos"
        ],
        "summary": "删除已上传视频",
        "operationId": "delete_video_api_videos__video_id__delete",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_video_api_videos_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_video_api_videos_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "path",
          "size_bytes",
          "duration_sec",
          "width",
          "height",
          "created_at"
        ],
        "title": "VideoOut"
      },
...
    }
  },
...
}